refactor(commands): split cli install parsing
